Add bot panel references for brains
