+++ Assets/Scripts/Systems/Save/FacilityArrayNormalizer.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: c409dfc1812f49dead2f30df5df597ab
+
