+++ Assets/Scripts/Systems/Save/SaveLoadCandidateSelector.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a0f7de482ed1422c84c9e3a7c8aec693
+
