+++ Assets/Scripts/Systems/Save/SaveSnapshotBuilder.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: bc2fa08cb9174b5092b5135855b3ee15
+
